+++ src/components/index.js
@@ -1,16 +1,16 @@
 export { default as AccountNav } from './account/AccountNav';
-export { default as ChatWindow } from './movieDisplays/ChatWindow';
+export { default as ChatWindow } from './movie-displays/ChatWindow';
 export { default as FeedbackToast } from './FeedbackToast';
-export { default as GenreList } from './movieDisplays/GenreList';
-export { default as GoogleLogin } from './loginAndSignup/GoogleLogin';
-export { default as Login } from './loginAndSignup/Login';
-export { default as MovieCard } from './movieDisplays/MovieCard';
-export { default as MovieFeed } from './movieDisplays/MovieFeed';
-export { default as MovieFeedItem } from './movieDisplays/MovieFeedItem';
-export { default as MovieList } from './movieDisplays/MovieList';
+export { default as GenreList } from './movie-displays/GenreList';
+export { default as GoogleLogin } from './login-signup/GoogleLogin';
+export { default as Login } from './login-signup/Login';
+export { default as MovieCard } from './movie-displays/MovieCard';
+export { default as MovieFeed } from './movie-displays/MovieFeed';
+export { default as MovieFeedItem } from './movie-displays/MovieFeedItem';
+export { default as MovieList } from './movie-displays/MovieList';
 export { default as Navbar } from './Navbar';
-export { default as OAuth } from './loginAndSignup/OAuth';
-export { default as SearchBar } from './movieDisplays/SearchBar';
-export { default as Signup } from './loginAndSignup/Signup';
-export { default as TrailerModal } from './movieDisplays/TrailerModal';
+export { default as OAuth } from './login-signup/OAuth';
+export { default as SearchBar } from './movie-displays/SearchBar';
+export { default as Signup } from './login-signup/Signup';
+export { default as TrailerModal } from './movie-displays/TrailerModal';
 export { default as Watchlist } from './account/Watchlist';
