+++ src/components/login-signup/GoogleLogin.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import '../../css/loginAndSignup/GoogleLogin.css';
+import '../../css/login-signup/GoogleLogin.css';
 
 const GoogleLogin = () => (
   <a href={`${process.env.REACT_APP_SVR_AUTH}/google`} className="google-link">
