+++ src/css/movie-displays/MovieCard.css
