+++ src/css/movie-displays/SearchBar.css
