+++ src/pages/UserDashboardPage.js
@@ -1,7 +1,7 @@
 /* eslint-disable react/prefer-stateless-function */
 import React, { Component } from 'react';
 import '../css/main.css';
-import '../css/cssPages/UserDashboardPage.css';
+import '../css/css-pages/UserDashboardPage.css';
 import { Route } from 'react-router-dom';
 import AccountNav from '../components/account/AccountNav';
 import UserPreferences from '../components/account/UserPreferences';
